Fixed ActivityPub feed link issues
ref AP-908

- Links in feed items should always open in a new tab. ATM hashtags and mentions open in same window
- Clicking on a link shouldn't open the drawer
