+++ apps/admin-x-activitypub/src/components/feed/ArticleModal.tsx
@@ -21,6 +21,7 @@ import DeletedFeedItem from './DeletedFeedItem';
 import TableOfContents, {TOCItem} from './TableOfContents';
 import getReadingTime from '../../utils/get-reading-time';
 import {isPendingActivity} from '../../utils/pending-activity';
+import {openLinksInNewTab} from '@src/utils/content-formatters';
 import {useDebounce} from 'use-debounce';
 
 interface ArticleModalProps {
@@ -75,23 +76,6 @@ const ArticleBody: React.FC<{
     const darkMode = document.documentElement.classList.contains('dark');
 
     const cssContent = articleBodyStyles(siteData?.url.replace(/\/$/, ''));
-
-    const openLinksInNewTab = (content: string) => {
-        // Create a temporary div to parse the HTML
-        const div = document.createElement('div');
-        div.innerHTML = content;
-
-        // Find all anchor tags
-        const links = div.getElementsByTagName('a');
-
-        // Add target="_blank" and rel attributes to each link
-        for (let i = 0; i < links.length; i++) {
-            links[i].setAttribute('target', '_blank');
-            links[i].setAttribute('rel', 'noopener noreferrer');
-        }
-
-        return div.innerHTML;
-    };
 
     const htmlContent = `
         <html class="has-${!darkMode ? 'dark' : 'light'}-text">
