+++ apps/admin-x-activitypub/src/utils/content-formatters.ts
@@ -1,0 +1,20 @@
+export function stripHtml(html: string): string {
+    return html.replace(/<[^>]*>/g, '');
+}
+
+export const openLinksInNewTab = (content: string) => {
+    // Create a temporary div to parse the HTML
+    const div = document.createElement('div');
+    div.innerHTML = content;
+
+    // Find all anchor tags
+    const links = div.getElementsByTagName('a');
+
+    // Add target="_blank" and rel attributes to each link
+    for (let i = 0; i < links.length; i++) {
+        links[i].setAttribute('target', '_blank');
+        links[i].setAttribute('rel', 'noopener noreferrer');
+    }
+
+    return div.innerHTML;
+};
