--- apps/admin-x-activitypub/src/utils/strip-html.ts
@@ -1,3 +1,0 @@
-export default function stripHtml(html: string): string {
-    return html.replace(/<[^>]*>/g, '');
-}
